EXAMSYS-3381 Stop errors when module is not in ExamSys
When a module we get enrolment instructions on cannot be found in ExamSys false is returned.

Before this change we were then still trying to access this boolean as though it was an array to find the internal id of the module.

After this change we immediately check if a module was found.

While making this change it was possible to refactor a layer of nesting out of the code, which will hopefully make it easier to follow.