EXAMSYS-3393 Handle non-targeted enrolment calls
Before this change there would be issues when trying to get all enrolments without the targeted enrolments setting being enabled.

This happened because the code assumed that there was an external id, when there was not an invalid value was passed to the get academic year start  method, which caused issues.